Clarified Battleship-G1, fixed setup() split 1. Added more clarification comments of what executive methods do, to make clear that setup() can set up the board for each player. (Also, fixed this, so not split into 2 methods.) This also fixed each player being asked number of ships (it only needs to be asked once, before the 2 payer's boards asre set up).
2. Moved Player1 and Player2 string variables to the playGame method because that is
where they will be used, probably. Can move out if they need to have global scope.
They seem like they would be useful but it depends on how playGame() is designed.

3. Fixed parameters for playGame (the inputs are the two Board objects after
modification by the setup method.

4. Did not alter the printMenu method but wrote a better explanation for what this
is needed for. Most important is letting player choose between proceeding with the
game and taking a shot, or quitting, at every single turn. I mostly liked it as is,
especially that it just gives the option to print instructions in case needed, rather
than printing them by default.

5. I removed the #'s that were added to my print statements and undid their removal,
since the welocme statement would only need to be printed a single time at the start
of the game, and do not need to be in a printMenu that is called repeatedly between
player turns. I also had to put them back since number of ships only needs to be asked
a single time for a game, so the print statements were correct.

A minor fix:
Python convention (and I think an option that Doxygen recognizes, to help make the
documentation) is to have """explanation""" (or three single quotes on each side)
beneath the method declaration line, so I moved them back there because some had been
moved beneath definition bloc. For now some of these are temporary, to aid our work,
but probably nice to leave clear explanation for doxygen (or Sphinx; maybe that will work too).
Probably do not delete the temporary explanations 'till we are done, to help everyone
understand the goals of the method, and what needs to be fixed or added to the method.
But do edit these, as needed. Just don't remove, at least not yet.